adhere to Arduino IDE standards for sketch names (#65)
The IDE wants them to be named like their containing folder. Otherwise it proposes to create a new
folder and to move the sketch into it, which again causes troubles because the header files are not
automatically moved. Giving the sketch name the proper name avoids a lot of troubles for Arduino
IDE users (at least for version 2.3.4)

I missed this due to bad practice on my part, and this problem is found with all versions of Arduino IDE that I know of.